cloud/requirements.txt
pyserial
paho-mqtt
requests
flask

